yet more warning fixes.

+++ tests/relay_test.c
@@ -265,7 +265,7 @@ size_t nb_relay_test_groups_objects[3] = { 4, 4, 1 };
 int quicrq_relay_cache_verify(quicrq_relay_cached_media_t* cached_ctx)
 {
     int ret = 0;
-    size_t nb_fragments_found = 0;
+    int nb_fragments_found = 0;
     for (size_t f_id = 0; ret == 0 && f_id < nb_relay_test_objects; f_id++) {
         size_t offset = 0;
         while (ret == 0 && offset < relay_test_objects[f_id].length) {
@@ -310,7 +310,7 @@ int quicrq_relay_cache_verify(quicrq_relay_cached_media_t* cached_ctx)
     if (ret == 0) {
         /* Check that cache contains exactly the expected number of fragments */
         if (cached_ctx->fragment_tree.size != nb_fragments_found) {
-            DBG_PRINTF("Found %zu fragments, cache contains %zu", nb_fragments_found, cached_ctx->fragment_tree.size);
+            DBG_PRINTF("Found %d fragments, cache contains %d", nb_fragments_found, cached_ctx->fragment_tree.size);
             ret = -1;
         }
     }
@@ -318,14 +318,14 @@ int quicrq_relay_cache_verify(quicrq_relay_cached_media_t* cached_ctx)
         /* Verify the chain of fragments */
         quicrq_relay_cached_fragment_t* fragment = cached_ctx->first_fragment;
         quicrq_relay_cached_fragment_t* previous_fragment = NULL;
-        size_t nb_in_chain = 0;
+        int nb_in_chain = 0;
         while (fragment != NULL) {
             nb_in_chain++;
             previous_fragment = fragment;
             fragment = fragment->next_in_order;
         }
         if (nb_in_chain != cached_ctx->fragment_tree.size) {
-            DBG_PRINTF("Found %zu fragments in chain, cache contains %zu", nb_in_chain, cached_ctx->fragment_tree.size);
+            DBG_PRINTF("Found %d fragments in chain, cache contains %d", nb_in_chain, cached_ctx->fragment_tree.size);
             ret = -1;
         }
         else if (previous_fragment != cached_ctx->last_fragment) {
